rearrange cpio output format selection
