+++ libarchive/test/test_write_format_cpio.c
@@ -276,12 +276,12 @@ DEFINE_TEST(test_write_format_cpio)
 	int64_t size_4g = ((int64_t)1) << 32;
 	int64_t size_8g = ((int64_t)1) << 33;
 
-	test_format(archive_write_set_format_cpio);
+	test_format(archive_write_set_format_cpio_odc);
 	test_format(archive_write_set_format_cpio_newc);
 
-	test_big_entries(archive_write_set_format_cpio,
+	test_big_entries(archive_write_set_format_cpio_odc,
 	    size_8g - 1, ARCHIVE_OK);
-	test_big_entries(archive_write_set_format_cpio,
+	test_big_entries(archive_write_set_format_cpio_odc,
 	    size_8g, ARCHIVE_FAILED);
 	test_big_entries(archive_write_set_format_cpio_newc,
 	    size_4g - 1, ARCHIVE_OK);
