+++ libarchive/test/test_write_format_cpio_empty.c
@@ -52,7 +52,7 @@ DEFINE_TEST(test_write_format_cpio_empty)
 
 	/* Create a new archive in memory. */
 	assert((a = archive_write_new()) != NULL);
-	assertA(0 == archive_write_set_format_cpio(a));
+	assertA(0 == archive_write_set_format_cpio_odc(a));
 	assertA(0 == archive_write_add_filter_none(a));
 	/* 1-byte block size ensures we see only the required bytes. */
 	/* We're not testing the padding here. */
